pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-board"
version = "0.1.0"
description = "Personal job board aggregator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "numpy",
    "requests",
    "openai",
    "pydantic",
    "playwright",
    "beautifulsoup4",
    "sqlalchemy",
    "tqdm",
    "streamlit",
    "tabulate",
    "PyYAML"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
